Models/requirements.txt
langchain
langchain-core
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-huggingface
numpy
scikit-learn
transformers
huggingface_hub
python-dotenv
TensorFlow
#pip install -r requirements.txt  